update packaging
it was just adding 4 characters to every import path
