+++ examples/encoder/README.md
@@ -2,13 +2,13 @@
 
 # Build & Run
 
-```
+```shell
 go build -o bin/ .
 ./bin/encoder base64encode str1 | xargs -L1 ./bin/encoder base64decode
 ```
 
 # Run
 
-```
+```shell
 go run . base64encode str1 
 ```
