+++ examples/mockserver/README.md
@@ -2,14 +2,13 @@
 
 # Build & Run
 
-```
+```shell
 go build -o bin/ .
 ./bin/mockserver
 ```
 
 # Run
 
-
-```
+```shell
 go run .
 ```
