+++ examples/mockserver/main.go
@@ -3,12 +3,12 @@ package main
 import (
 	"fmt"
 	"github.com/fsnotify/fsnotify"
-	"github.com/lnashier/go-app/pkg/config"
-	"github.com/lnashier/go-app/pkg/env"
-	chandler "github.com/lnashier/go-app/pkg/http/handler"
-	"github.com/lnashier/go-app/pkg/http/service"
-	"github.com/lnashier/go-app/pkg/log"
-	"github.com/lnashier/go-app/pkg/zson"
+	"github.com/lnashier/go-app/config"
+	"github.com/lnashier/go-app/env"
+	chandler "github.com/lnashier/go-app/http/handler"
+	"github.com/lnashier/go-app/http/service"
+	"github.com/lnashier/go-app/log"
+	"github.com/lnashier/go-app/zson"
 	"net/http"
 )
