+++ examples/myhttpapp/README.md
@@ -2,13 +2,13 @@
 
 # Build & Run
 
-```
+```shell
 go build -o ./bin/myhttpapp ./cmd
 ./bin/myhttpapp
 ```
 
 # Run
 
-```
+```shell
 go run ./cmd
 ```
