+++ examples/myhttpapp/cmd/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
+	"github.com/lnashier/go-app/config"
 	"github.com/lnashier/go-app/examples/myhttpapp/app"
-	"github.com/lnashier/go-app/pkg/config"
-	"github.com/lnashier/go-app/pkg/http/service"
+	"github.com/lnashier/go-app/http/service"
 )
 
 func main() {
