+++ http/handler/xml.go
