+++ http/service/server.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"errors"
 	"github.com/gorilla/mux"
-	"github.com/lnashier/go-app/pkg/buildinfo"
-	"github.com/lnashier/go-app/pkg/config"
-	"github.com/lnashier/go-app/pkg/log"
+	"github.com/lnashier/go-app/buildinfo"
+	"github.com/lnashier/go-app/config"
+	"github.com/lnashier/go-app/log"
 	"github.com/urfave/negroni"
 	"net/http"
 	"strings"
